The client now uses Meteor cursor.observeChanges() in the collection "method_commands", and captures the modifications in this collection after the Event Receiver nodeja app receives a request for an EventResponse async message from the sila device/simulator.
PERFORMANCE BUG!! There is a timeout for commands called in parallel to the simulator. This is a workaround for the bug in HSR bug for not allowing to call multiple commands in parallel (try calling getStatus at the same time sequentially and it fails). This timeout gives some interval of time so the simulator can process only one command at the time. Delete this workaround for a better simulator or a device that allows parallel command calls/execution.